:rotating_light: condensed functionality for issue #99
