+++ src/Entities/Filesystem/FileIgnored.php
@@ -2,7 +2,6 @@
 
 namespace Tapestry\Entities\Filesystem;
 
-use Tapestry\Entities\File;
 use Symfony\Component\Filesystem\Filesystem;
 use Symfony\Component\Console\Output\OutputInterface;
 
@@ -13,28 +12,14 @@ use Symfony\Component\Console\Output\OutputInterface;
  * files that have not changed since the last time Tapestry ran - this is so that Tapestry only writes files that need
  * writing and doesn't crash browser sync.
  */
-class FileIgnored implements FilesystemInterface
+class FileIgnored extends FileAction implements FilesystemInterface
 {
     /**
-     * @var File
+     * @param Filesystem      $filesystem
+     * @param OutputInterface $output
+     *
+     * @return void
      */
-    private $file;
-    private $destinationPath;
-
-    public function __construct(File $file, $destinationPath)
-    {
-        $this->file = $file;
-        $this->destinationPath = $destinationPath;
-    }
-
-    /**
-     * @return File
-     */
-    public function getFile()
-    {
-        return $this->file;
-    }
-
     public function __invoke(Filesystem $filesystem, OutputInterface $output)
     {
         $output->writeln('[+] Ignoring File ['.$this->file->getUid().']');
